1、新增9宫格控件

+++ uiwidget/src/main/java/com/hearthappy/uiwidget/utils/PxUtils.kt
@@ -1,6 +1,5 @@
 package com.hearthappy.uiwidget.utils
 
-import android.content.Context
 import android.content.res.Resources
 
 
@@ -11,24 +10,24 @@ import android.content.res.Resources
  *
  */
 
-fun Int.sp2px(): Float {
-    return this * Resources.getSystem().displayMetrics.scaledDensity + 0.5f
+fun Int.sp2px(): Int {
+    return (this * Resources.getSystem().displayMetrics.scaledDensity + 0.5f).toInt()
 }
 
 fun Float.sp2px(): Float {
     return this * Resources.getSystem().displayMetrics.scaledDensity + 0.5f
 }
 
-fun Int.dp2px(): Float {
-    return Resources.getSystem().displayMetrics.density * this
+fun Int.dp2px(): Int {
+    return (Resources.getSystem().displayMetrics.density * this).toInt()
 }
 
 fun Float.dp2px(): Float {
     return Resources.getSystem().displayMetrics.density * this
 }
 
-fun Int.px2dp(): Float {
-    return (this / Resources.getSystem().displayMetrics.density)
+fun Int.px2dp(): Int {
+    return ((this / Resources.getSystem().displayMetrics.density).toInt())
 
 }
